Reshuffle - separate yaml and ui config moudles
